+++ src/test/customApi/odataOperationRegistry.test.ts
@@ -1,5 +1,4 @@
 import assert from "node:assert/strict";
-import { suite, test } from "mocha";
 import { parseODataOperationRegistry } from "../../customApi/odata/odataMetadataParser.js";
 import { applyCustomApiExecutionEligibility, resolveCustomApiExecutionEligibility } from "../../customApi/odata/odataOperationEligibility.js";
 import type { CustomApiDefinition } from "../../customApi/models/customApiTypes.js";